change ~ use Output structure in place of global `atomics` for vertical tracking
- using output tracking (`colors_shown`) instead of `first_output`
  - the `colors_shown` property is more concrete and is more "correct" no matter where/when Output is used
