+++ src/cli/commands/gray.rs
@@ -1,5 +1,7 @@
+use std::io::Write;
+
 use crate::commands::prelude::*;
-use crate::commands::show::show_color;
+use crate::output::Output;
 
 use pastel::Color;
 
@@ -7,8 +9,9 @@ pub struct GrayCommand;
 
 impl GenericCommand for GrayCommand {
     fn run(&self, out: &mut dyn Write, matches: &ArgMatches, config: &Config) -> Result<()> {
+        let mut o = Output::new(out);
         let lightness = number_arg(matches, "lightness")?;
         let gray = Color::graytone(lightness);
-        show_color(out, &config, &gray)
+        o.show_color(&config, &gray)
     }
 }
