+++ src/cli/commands/random.rs
@@ -1,5 +1,7 @@
+use std::io::Write;
+
 use crate::commands::prelude::*;
-use crate::commands::show::show_color;
+use crate::output::Output;
 
 use pastel::random::strategies;
 use pastel::random::RandomizationStrategy;
@@ -8,6 +10,7 @@ pub struct RandomCommand;
 
 impl GenericCommand for RandomCommand {
     fn run(&self, out: &mut dyn Write, matches: &ArgMatches, config: &Config) -> Result<()> {
+        let mut o = Output::new(out);
         let strategy_arg = matches.value_of("strategy").expect("required argument");
 
         let count = matches.value_of("number").expect("required argument");
@@ -24,7 +27,7 @@ impl GenericCommand for RandomCommand {
         };
 
         for _ in 0..count {
-            show_color(out, &config, &strategy.generate())?;
+            o.show_color(&config, &strategy.generate())?;
         }
 
         Ok(())
