+++ src/cli/main.rs
@@ -10,6 +10,7 @@ mod config;
 mod error;
 mod hdcanvas;
 mod named;
+mod output;
 mod parser;
 mod utility;
